azure-common-resources/create-auto-restart/functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-mgmt-containerservice
azure-mgmt-kusto
azure-identity
azure-mgmt-resource
azure-mgmt-compute
holidays